qtdemux: Correctly parse mvhd atoms
Parse mvhd data according to its version to avoid failing
on valid files.

+++ gst/qtdemux/qtdemux.h
@@ -70,7 +70,7 @@ struct _GstQTDemux {
   GNode *moov_node_compressed;
 
   guint32 timescale;
-  guint32 duration;
+  guint64 duration;
 
   gint state;
